refactored calculateWithdrawal into two functions
balance available to be withdrawn is now shown by streams
